Cleaned up code and examples, renamed unit tests, added File.Delete calls
